+++ src/symbols/Remove/Remove.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useState } from 'react';
 import classNames from 'classnames';
-import { motion } from 'framer-motion';
+import { motion } from 'framer-motion/dist/framer-motion';
 import css from './Remove.module.css';
 
 export interface RemoveProps {
@@ -46,7 +46,7 @@ export const Remove: FC<Partial<RemoveProps>> = ({
         className={css.drop}
         onMouseEnter={onEnter}
         onMouseLeave={onLeave}
-        onClick={(event) => {
+        onClick={event => {
           event.preventDefault();
           event.stopPropagation();
           onClick(event);