+++ stories/Editor.stories.tsx
@@ -1,7 +1,7 @@
 import React, { useRef, useState } from 'react';
 import { Canvas, CanvasRef } from '../src/Canvas';
 import { Node, Edge, MarkerArrow, Port, Icon, Arrow, Label, Remove, Add } from '../src/symbols';
-import { motion, useDragControls } from 'framer-motion';
+import { motion, useDragControls } from 'framer-motion/dist/framer-motion';
 import { Portal } from 'rdk';
 import { EdgeData, NodeData } from '../src/types';
 import { addNodeAndEdge, useProximity } from '../src/helpers';